[FIX] project_timeline: Allow to install this module with project_enterprise
This PR fixes an issue added in commit
https://github.com/OCA/project/commit/3689f1dfccf755c5d862c0e779ab0364215dea3a

When the module project_timeline is installed in a database with the module
project_enterprise, it raises the following error:

psycopg2.errors.DuplicateColumn: column "planned_date_end" specified more than once

This is because the field planned_date_end is defined in both modules.

This PR checks if the field is already defined in the query before adding it.